WIP: Hack in changes that prevent NSFW-related crashes
Prior to these changes, running the `nsfw` stress tests with the `nsfw` 
repo open in Atom will reliably crash the main process and render 
process within a few minutes. So far, with these changes, I haven't been 
able to reproduce a crash via the same method. We obviously can't ship 
with all of these changes as they are, but I wanted to at least record 
the finding.